WIP float16 support

+++ src/model.py
@@ -40,8 +40,8 @@ def norm(x, scope, *, axis=-1, epsilon=1e-5, hparams=None):
     dtype = hparams.dtype if hparams else tf.float32
     with tf.variable_scope(scope, dtype=dtype):
         n_state = x.shape[-1].value
-        g = get_variable('g') or tf.get_variable('g', [n_state], initializer=tf.constant_initializer(1))
-        b = get_variable('b') or tf.get_variable('b', [n_state], initializer=tf.constant_initializer(0))
+        g = get_variable('g') or tf.get_variable('g', [n_state], initializer=tf.constant_initializer(1, dtype=dtype))
+        b = get_variable('b') or tf.get_variable('b', [n_state], initializer=tf.constant_initializer(0, dtype=dtype))
         u = tf.reduce_mean(x, axis=axis, keepdims=True)
         s = tf.reduce_mean(tf.square(x-u), axis=axis, keepdims=True)
         x = (x - u) * tf.rsqrt(s + epsilon)
@@ -62,8 +62,8 @@ def conv1d(x, scope, nf, *, w_init_stdev=0.02, hparams=None):
     dtype = hparams.dtype if hparams else tf.float32
     with tf.variable_scope(scope, dtype=dtype):
         *start, nx = shape_list(x)
-        w = get_variable('w') or tf.get_variable('w', [1, nx, nf], initializer=tf.random_normal_initializer(stddev=w_init_stdev))
-        b = get_variable('b') or tf.get_variable('b', [nf], initializer=tf.constant_initializer(0))
+        w = get_variable('w') or tf.get_variable('w', [1, nx, nf], initializer=tf.random_normal_initializer(stddev=w_init_stdev, dtype=dtype))
+        b = get_variable('b') or tf.get_variable('b', [nf], initializer=tf.constant_initializer(0, dtype=dtype))
         c = tf.reshape(tf.matmul(tf.reshape(x, [-1, nx]), tf.reshape(w, [-1, nf]))+b, start+[nf])
         return c
 
@@ -166,9 +166,9 @@ def model(hparams, X, past=None, scope='model', reuse=tf.AUTO_REUSE):
         batch, sequence = shape_list(X)
 
         wpe = get_variable('wpe') or tf.get_variable('wpe', [hparams.n_ctx, hparams.n_embd],
-                             initializer=tf.random_normal_initializer(stddev=0.01))
+                             initializer=tf.random_normal_initializer(stddev=0.01, dtype=dtype))
         wte = get_variable('wte') or tf.get_variable('wte', [hparams.n_vocab, hparams.n_embd],
-                             initializer=tf.random_normal_initializer(stddev=0.02))
+                             initializer=tf.random_normal_initializer(stddev=0.02, dtype=dtype))
         past_length = 0 if past is None else tf.shape(past)[-2]
         h = tf.gather(wte, X) + tf.gather(wpe, positions_for(X, past_length))
